+++ src/command_modules/azure-cli-cosmosdb/azure/cli/command_modules/cosmosdb/tests/latest/test_cosmosdb_commands.py
@@ -4,6 +4,7 @@
 # --------------------------------------------------------------------------------------------
 
 from azure.cli.testsdk import JMESPathCheck, ScenarioTest, ResourceGroupPreparer
+from knack.util import CLIError
 
 
 class CosmosDBTests(ScenarioTest):
@@ -152,8 +153,7 @@ class CosmosDBTests(ScenarioTest):
         account2 = self.cmd('az cosmosdb show -n {acc} -g {rg}').get_output_in_json()
         assert len(account2['writeLocations']) == 1
         assert len(account2['readLocations']) == 2
-        print(account2['writeLocations'][0]['failoverPriority'])
-        print(account2['writeLocations'][0]['locationName'])
+
         assert account2['writeLocations'][0]['failoverPriority'] == 0
         assert account2['writeLocations'][0]['locationName'] == "West US"
         assert account2['readLocations'][0]['locationName'] == "East US" or account2['readLocations'][1]['locationName'] == "East US"
@@ -191,3 +191,82 @@ class CosmosDBTests(ScenarioTest):
         self.cmd('az cosmosdb database list -n {acc} -g {rg}')
         self.cmd('az cosmosdb database list -n {acc} --key {primary_master_key}')
         self.cmd('az cosmosdb database list --url-connection {url} --key {primary_master_key}')
+
+    @ResourceGroupPreparer(name_prefix='cli_test_cosmosdb_account')
+    def test_cosmosdb_network_rule_list(self, resource_group):
+
+        self.kwargs.update({
+            'acc': self.create_random_name(prefix='cli', length=40),
+            'vnet': self.create_random_name(prefix='cli', length=40),
+            'sub': self.create_random_name(prefix='cli', length=40)
+        })
+
+        vnet_output = self.cmd('az network vnet create --name {vnet} --resource-group {rg} --subnet-name {sub}').get_output_in_json()
+        self.cmd('az network vnet subnet update -g {rg} --vnet-name {vnet} -n {sub} --service-endpoints Microsoft.AzureCosmosDB')
+
+        self.kwargs.update({
+            'subnet_id': vnet_output["newVNet"]["subnets"][0]["id"]
+        })
+
+        self.cmd('az cosmosdb create -n {acc} -g {rg} --enable-virtual-network --virtual-network-rule {subnet_id}').get_output_in_json()
+
+        vnet_rules = self.cmd('az cosmosdb network-rule list -n {acc} -g {rg}').get_output_in_json()
+
+        assert len(vnet_rules) == 1
+        assert vnet_rules[0]["id"] == vnet_output["newVNet"]["subnets"][0]["id"]
+
+    @ResourceGroupPreparer(name_prefix='cli_test_cosmosdb_account')
+    def test_cosmosdb_network_rule_add(self, resource_group):
+
+        self.kwargs.update({
+            'acc': self.create_random_name(prefix='cli', length=40),
+            'vnet': self.create_random_name(prefix='cli', length=40),
+            'sub': self.create_random_name(prefix='cli', length=40)
+        })
+
+        vnet_output = self.cmd('az network vnet create --name {vnet} --resource-group {rg} --subnet-name {sub}').get_output_in_json()
+
+        self.kwargs.update({
+            'subnet_id': vnet_output["newVNet"]["subnets"][0]["id"]
+        })
+
+        self.cmd('az cosmosdb create -n {acc} -g {rg} --enable-virtual-network')
+
+        with self.assertRaisesRegexp(CLIError, "usage error: --subnet ID | --subnet NAME --vnet-name NAME"):
+            self.cmd('az cosmosdb network-rule add -n {acc} -g {rg} --subnet {vnet}')
+
+        vnet_rule = self.cmd('az cosmosdb network-rule add -n {acc} -g {rg} --virtual-network {vnet} --subnet {sub} --ignore-missing-vnet-service-endpoint').get_output_in_json()
+
+        assert vnet_rule["virtualNetworkRules"][0]["id"] == vnet_output["newVNet"]["subnets"][0]["id"]
+        assert vnet_rule["virtualNetworkRules"][0]["ignoreMissingVnetServiceEndpoint"]
+
+        existing_rule = self.cmd('az cosmosdb network-rule add -n {acc} -g {rg} --virtual-network {vnet} --subnet {sub} --ignore-missing-vnet-service-endpoint').get_output_in_json()
+        assert len(existing_rule["virtualNetworkRules"]) == 1
+
+    @ResourceGroupPreparer(name_prefix='cli_test_cosmosdb_account')
+    def test_cosmosdb_network_rule_remove(self, resource_group):
+
+        self.kwargs.update({
+            'acc': self.create_random_name(prefix='cli', length=40),
+            'vnet': self.create_random_name(prefix='cli', length=40),
+            'sub': self.create_random_name(prefix='cli', length=40)
+        })
+
+        vnet_output = self.cmd('az network vnet create --name {vnet} --resource-group {rg} --subnet-name {sub}').get_output_in_json()
+
+        self.kwargs.update({
+            'subnet_id': vnet_output["newVNet"]["subnets"][0]["id"]
+        })
+
+        self.cmd('az cosmosdb create -n {acc} -g {rg} --enable-virtual-network')
+
+        vnet_rule = self.cmd('az cosmosdb network-rule add -n {acc} -g {rg} --subnet {subnet_id} --ignore-missing-vnet-service-endpoint').get_output_in_json()
+
+        assert vnet_rule["virtualNetworkRules"][0]["id"] == vnet_output["newVNet"]["subnets"][0]["id"]
+        assert vnet_rule["virtualNetworkRules"][0]["ignoreMissingVnetServiceEndpoint"]
+
+        self.cmd('az cosmosdb network-rule remove -n {acc} -g {rg} --subnet {subnet_id}')
+
+        vnet_rules = self.cmd('az cosmosdb network-rule list -n {acc} -g {rg}').get_output_in_json()
+
+        assert len(vnet_rules) == 0