+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_validators.py
@@ -3,12 +3,15 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
+# pylint: disable=too-many-lines
+
 import argparse
 import base64
 import socket
 import os
 
 from knack.util import CLIError
+from knack.log import get_logger
 
 from azure.cli.core.commands.validators import \
     (validate_tags, get_default_location_from_resource_group)
@@ -17,8 +20,8 @@ from azure.cli.core.commands.client_factory import get_subscription_id, get_mgmt
 from azure.cli.core.commands.validators import validate_parameter_set
 from azure.cli.core.profiles import ResourceType
 
-# PARAMETER VALIDATORS
-# pylint: disable=too-many-lines
+
+logger = get_logger(__name__)
 
 
 def get_asg_validator(loader, dest):
@@ -249,6 +252,93 @@ def validate_dns_record_type(namespace):
             return
 
 
+def validate_express_route_peering(cmd, namespace):
+    from msrestazure.tools import is_valid_resource_id, resource_id
+    circuit = namespace.circuit_name
+    peering = namespace.peering
+
+    if not circuit and not peering:
+        return
+
+    usage_error = CLIError('usage error: --peering ID | --peering NAME --circuit-name CIRCUIT')
+    if not is_valid_resource_id(peering):
+        namespace.peering = resource_id(
+            subscription=get_subscription_id(cmd.cli_ctx),
+            resource_group=namespace.resource_group_name,
+            namespace='Microsoft.Network',
+            type='expressRouteCircuits',
+            name=circuit,
+            child_type_1='peerings',
+            child_name_1=peering
+        )
+    elif circuit:
+        raise usage_error
+
+
+def validate_express_route_port(cmd, namespace):
+    from msrestazure.tools import is_valid_resource_id, resource_id
+    if namespace.express_route_port and not is_valid_resource_id(namespace.express_route_port):
+        namespace.express_route_port = resource_id(
+            subscription=get_subscription_id(cmd.cli_ctx),
+            resource_group=namespace.resource_group_name,
+            namespace='Microsoft.Network',
+            type='expressRoutePorts',
+            name=namespace.express_route_port
+        )
+
+
+def validate_virtual_hub(cmd, namespace):
+    from msrestazure.tools import is_valid_resource_id, resource_id
+    if namespace.virtual_hub and not is_valid_resource_id(namespace.virtual_hub):
+        namespace.virtual_hub = resource_id(
+            subscription=get_subscription_id(cmd.cli_ctx),
+            resource_group=namespace.resource_group_name,
+            namespace='Microsoft.Network',
+            type='virtualHubs',
+            name=namespace.virtual_hub
+        )
+
+
+def bandwidth_validator_factory(mbps=True):
+    def validator(namespace):
+        return validate_circuit_bandwidth(namespace, mbps=mbps)
+    return validator
+
+
+def validate_circuit_bandwidth(namespace, mbps=True):
+    # use gbps if mbps is False
+    unit = 'mbps' if mbps else 'gbps'
+    bandwidth = None
+    bandwidth = getattr(namespace, 'bandwidth_in_{}'.format(unit), None)
+    if bandwidth is None:
+        return
+
+    if len(bandwidth) == 1:
+        bandwidth_comps = bandwidth[0].split(' ')
+    else:
+        bandwidth_comps = bandwidth
+
+    usage_error = CLIError('usage error: --bandwidth INT {Mbps,Gbps}')
+    if len(bandwidth_comps) == 1:
+        logger.warning('interpretting --bandwidth as %s. Consider being explicit: Mbps, Gbps', unit)
+        setattr(namespace, 'bandwidth_in_{}'.format(unit), float(bandwidth_comps[0]))
+        return
+    if len(bandwidth_comps) > 2:
+        raise usage_error
+
+    if float(bandwidth_comps[0]) and bandwidth_comps[1].lower() in ['mbps', 'gbps']:
+        input_unit = bandwidth_comps[1].lower()
+        if input_unit == unit:
+            converted_bandwidth = float(bandwidth_comps[0])
+        elif input_unit == 'gbps':
+            converted_bandwidth = float(bandwidth_comps[0]) * 1000
+        else:
+            converted_bandwidth = float(bandwidth_comps[0]) / 1000
+        setattr(namespace, 'bandwidth_in_{}'.format(unit), converted_bandwidth)
+    else:
+        raise usage_error
+
+
 def validate_er_peer_circuit(cmd, namespace):
     from msrestazure.tools import resource_id, is_valid_resource_id
 
@@ -313,6 +403,17 @@ def validate_frontend_ip_configs(cmd, namespace):
             else:
                 config_ids.append(item)
         namespace.frontend_ip_configurations = config_ids
+
+
+def validate_local_gateway(cmd, namespace):
+    from msrestazure.tools import is_valid_resource_id, resource_id
+    if namespace.gateway_default_site and not is_valid_resource_id(namespace.gateway_default_site):
+        namespace.gateway_default_site = resource_id(
+            subscription=get_subscription_id(cmd.cli_ctx),
+            resource_group=namespace.resource_group_name,
+            name=namespace.gateway_default_site,
+            namespace='Microsoft.Network',
+            type='localNetworkGateways')
 
 
 def validate_metadata(namespace):
@@ -829,6 +930,8 @@ def process_vnet_gateway_create_namespace(cmd, namespace):
     if public_ip_count > 2:
         raise CLIError('Specify a single public IP to create an active-standby gateway or two '
                        'public IPs to create an active-active gateway.')
+
+    validate_local_gateway(cmd, ns)
 
     enable_bgp = any([ns.asn, ns.bgp_peering_address, ns.peer_weight])
     if enable_bgp and not ns.asn: