+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/tests/latest/test_graph.py
@@ -95,12 +95,27 @@ class ApplicationSetScenarioTest(ScenarioTest):
         name = self.create_random_name(prefix='cli-graph', length=14)
         self.kwargs.update({
             'app': 'http://' + name,
-            'name': name
+            'name': name,
+            'app_roles': json.dumps([
+                {
+                    "allowedMemberTypes": [
+                        "User"
+                    ],
+                    "description": "Approvers can mark documents as approved",
+                    "displayName": "Approver",
+                    "isEnabled": "true",
+                    "value": "approver"
+                }
+            ])
         })
 
-        # crerate app through general option
+        # create app through general option
         self.cmd('ad app create --display-name {name} --homepage {app} --identifier-uris {app}',
                  checks=self.check('identifierUris[0]', '{app}'))
+
+        # set app password
+        result = self.cmd('ad app credential reset --id {app} --append --password "test" --years 2').get_output_in_json()
+        self.assertTrue(result['appId'])
 
         # show/list app
         self.cmd('ad app show --id {app}',
@@ -120,6 +135,11 @@ class ApplicationSetScenarioTest(ScenarioTest):
         self.cmd('ad app update --id {app} --set oauth2AllowUrlPathMatching=true')
         self.cmd('ad app show --id {app}',
                  checks=self.check('oauth2AllowUrlPathMatching', True))
+
+        # update app_roles
+        self.cmd("ad app update --id {app} --app-roles '{app_roles}'")
+        self.cmd('ad app show --id {app}',
+                 checks=self.check('length(appRoles)', 1))
 
         # delete app
         self.cmd('ad app delete --id {app}')