+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -3,11 +3,16 @@
 Release History
 ===============
 
+2.2.1
++++++
+* Fix redundant sources in image import.
+* Add '--target' parameter for 'az acr build', 'az acr task create' and 'az acr task update' commands.
+* Improve error handling for runtime commands when not logged into Azure.
+
 2.2.0
 +++++
 * BREAKING CHANGE: Remove 'acr build-task' command group.
-* BREAKING CHANGE: Remove --tag/--manifest from 'acr repository delete' command.
-* Add '--target' parameter for 'az acr build', 'az acr task create' and 'az acr task update' commands.
+* BREAKING CHANGE: Remove '--tag'/'--manifest' from 'acr repository delete' command.
 
 2.1.13
 ++++++