+++ src/command_modules/azure-cli-cosmosdb/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+0.2.8
++++++
+* Introduce `network-rule` subgroup with commands `add`, `remove`, and `list` for managing VNET rules of a Cosmos DB account
+
 0.2.7
 +++++
 * Added support for creating database with shared throughput