+++ src/command_modules/azure-cli-iotcentral/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+0.1.6
++++++
+* Minor fixes
+
 0.1.5
 +++++
 * Fix update command API call.